LayoutTests/media/modern-media-controls/audio/audio-controls-styles.html

<!DOCTYPE html>
<script src="../../../resources/js-test-pre.js"></script>
<body>
<script type="text/javascript">

description("Testing media elements disable built-in browser UI features to copy content.");

const audio = document.body.appendChild(document.createElement("audio"));
audio.controls = true;

shouldBeEqualToString("getComputedStyle(audio).WebkitUserSelect", "none");
shouldBeEqualToString("getComputedStyle(audio).WebkitTouchCallout", "none");

debug("");
audio.remove();

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
