Safari-609.1.11.1/LayoutTests/media/modern-media-controls/audio/audio-controls-styles-expected.txt

Testing media elements disable built-in browser UI features to copy content.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(audio).WebkitUserSelect is "none"
PASS getComputedStyle(audio).WebkitTouchCallout is "none"

PASS successfullyParsed is true

TEST COMPLETE
